requirements.txt
openai
python-dotenv
pytelegrambotapi
